requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
dotenv
langchain_community
langchain_core
langchain-text-splitters
langchain
pypdf
transformers
torch
tiktoken
SentencePiece
protobuf
torchvision
torchaudio
pydantic
ollama
docx
odfpy
sentence_transformers
faiss-cpu
faiss-gpu
pytorch 
pytorch 
pytorch-cuda
einops
requests